#341 Refactor: rename modal product and notification product
