+++ react/practice-two/src/components/index.ts
@@ -13,4 +13,4 @@ export { default as TableCell } from './Table/TableCell';
 export { default as TableHeader } from './Table/TableHeader';
 export { default as TableRow } from './Table/TableRow';
 export { default as Typography } from './Typography';
-export { default as NotificationModal } from './Modal/NotificationModal';
+export { default as ModalNotification } from './Modal/ModalNotification';
