+++ react/practice-two/src/layouts/HomeLayout/index.tsx
@@ -4,10 +4,10 @@ import { ChangeEvent, useCallback, useContext, useEffect, useState } from 'react
 import './index.css';
 
 // Components
-import { SelectItemProps, NotificationModal } from '@components';
+import { SelectItemProps, ModalNotification } from '@components';
 
 // Components of page
-import { ProductsTable, ProductModal, DataProduct } from '@pages';
+import { ProductsTable, ModalProduct, DataProduct } from '@pages';
 
 // Services
 import { getTypes, getStatuses, deleteProduct, getProductsByParam } from '@services';
@@ -176,7 +176,7 @@ const HomeLayout = () => {
         handleSetProductItem={handleSetProductItem}
       />
       {itemModal && (
-        <ProductModal
+        <ModalProduct
           productItem={productItem}
           status={status}
           types={types}
@@ -184,7 +184,7 @@ const HomeLayout = () => {
         />
       )}
       {notificationModal && (
-        <NotificationModal
+        <ModalNotification
           id={productItem.id || ''}
           variant='confirm'
           description='Do you want to delete this ?'
@@ -194,7 +194,7 @@ const HomeLayout = () => {
         />
       )}
       {errorsModal.status && (
-        <NotificationModal
+        <ModalNotification
           variant='notification'
           description={errorsModal.message}
           onCancel={handleCancel}
