+++ react/practice-two/src/pages/index.ts
@@ -2,5 +2,5 @@ export { default as HomePage } from './HomePage';
 export * from './HomePage/components/TableProducts/ProductRow';
 export { default as ProductRow } from './HomePage/components/TableProducts/ProductRow';
 export { default as ActionMenu } from './HomePage/components/ActionMenu';
-export { default as ProductModal } from './HomePage/components/ProductModal';
+export { default as ModalProduct } from './HomePage/components/ModalProduct';
 export { default as ProductsTable } from './HomePage/components/TableProducts';
